added log of k_sapce
the lof of the k-space is now a referace that is sent to the Adversarial DNN in addition to the sampled data.
the sampled data that is sent to the sampler is in the form of the logarithm of the abs of the k-space.
that is be cause the values of the actual k-space are too small, so as a regularization term, we send the log og the abs of the k-space to sample from.
need to do: fix the log of the k_space to be per channel instead of the abs of the k_sapce